Shared:Selectors:People: rewrite to typescript and move from client
