+++ packages/client/src/components/dialogs/ChangePortalOwnerDialog/index.js
@@ -118,6 +118,7 @@ const ChangePortalOwnerDialog = ({
             onAccept={onAccept}
             onCancel={onBackClick}
             onBackClick={onBackClick}
+            currentUserId={id}
           />
         </ModalDialog.Container>
       )}
