+++ packages/shared/components/selector/index.tsx
@@ -1,1 +1,3 @@
 export { Selector } from "./Selector";
+
+export type { SelectorProps } from "./Selector.types";
