+++ packages/shared/selectors/People/PeopleSelector.styled.ts
