+++ packages/shared/utils/common.ts
@@ -26,7 +26,8 @@ import BackgroundPatternBlackReactSvgUrl from "PUBLIC_DIR/images/background.patt
 
 import { FolderType, RoomsType, ThemeKeys } from "../enums";
 import { LANGUAGE, RTL_LANGUAGES } from "../constants";
-import { TUser, TI18n } from "../types";
+import { TI18n } from "../types";
+import { TUser } from "../api/people/types";
 import { TFolder, TFile, TGetFolder } from "../api/files/types";
 import { TRoom } from "../api/rooms/types";
 import TopLoaderService from "../components/top-loading-indicator";
